requirements.txt
# requirements.txt

# Core LangChain & LangGraph components
langchain
langgraph
langchain-core
langchain-groq
langchain-community

#frontend
streamlit

# LLM APIs
openai  # in case of fallback
sentence-transformers
faiss-cpu

# Data processing
pandas
numpy
scikit-learn

# Networking & environment
requests
python-dotenv

# Logging and debugging
loguru

# Optional: Jupyter support for prototyping
jupyter
